fe-refactor: Alterando brilhos nas imagens, adicionando imagem do ambiente.
