--- frontend/src/pages/home/components/content/components/Cards.css
@@ -1,3 +1,0 @@
-.brightness-50-contrast-150 {
-    filter: brightness(55%) contrast(150%);
-  }
