+++ frontend/src/pages/home/components/content/components/Cards/Cards.css
@@ -1,0 +1,3 @@
+.brightness-50-contrast-150 {
+    filter: brightness(70%) contrast(150%);
+  }
